file should use zip entry for constructor
