+++ spec/xml_parser_spec.rb
@@ -1,25 +1,9 @@
 require "minitest/autorun"
 require "blacklight"
 require "pry"
+require_relative "mocks/mockzip"
 
 include Blacklight
-
-class MockZip
-  class MockEntry
-    attr_accessor(:name)
-    def initialize(name = "fake/path__xid-12.jpg")
-      @name = name
-    end
-  end
-
-  def initialize(entries = nil)
-    @entries = entries || [MockEntry.new, MockEntry.new, MockEntry.new]
-  end
-
-  def entries
-    @entries
-  end
-end
 
 describe Blacklight do
   describe "create_random_hex" do
@@ -188,11 +172,16 @@ describe Blacklight do
 
   describe "iterate_files" do
     it "should return array of files" do
-      result = Blacklight.iterate_files(MockZip.new)
-      assert_equal(result.size, 3)
+      mock_entries = [
+        MockZip::MockEntry.new("csfiles/home_dir/test__xid-12.jpg"),
+        MockZip::MockEntry.new("csfiles/home_dir/test__xid-12.jpg")
+      ]
+
+      result = Blacklight.iterate_files(MockZip.new(mock_entries))
+      assert_equal(result.size, 2)
       assert_equal(result.first.id, "xid-12")
-      assert_equal(result.first.name, "path__xid-12.jpg")
-      assert_equal(result.first.location, "fake/path__xid-12.jpg")
+      assert_equal(result.first.name, "test__xid-12.jpg")
+      assert_equal(result.first.location, "csfiles/home_dir/test__xid-12.jpg")
     end
   end
 end
